+++ src/main/java/com/example/WombatFm/User/UserController.java
@@ -1,6 +1,8 @@
 package com.example.WombatFm.User;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -50,5 +52,25 @@ public class UserController {
         }
 
         return "redirect:/login";
+    }
+
+    @PostMapping("/saveUserRole")
+    public String saveUserRole(@RequestParam(name = "uid") int userId, @RequestParam(name = "user_role") String role) {
+        this.userService.updateUserRole(userId, role);
+        return "redirect:/showUsers";
+    }
+
+    @PostMapping("/saveRoles")
+    public String saveRoles(@RequestParam(name = "roles") Map<String, String> roles) {
+        Map<Integer, String> userRoles = new HashMap<>();
+
+        for(Map.Entry<String, String> entry : roles.entrySet()) {
+            int userId = Integer.parseInt(entry.getKey());
+            String role = entry.getValue();
+            userRoles.put(userId, role);
+        }
+
+        userService.updateMultipleUserRole(userRoles);
+        return "redirect:/showUsers";
     }
 }
