+++ src/main/java/com/example/WombatFm/User/UserRepository.java
@@ -8,4 +8,5 @@ public interface UserRepository {
     Optional<User> findByUsername(String username);
     List<User> getAllUsersPaginated(int limit, int offset);
     int countUsers();
+    void updateUserRole(int userId, String role);
 }
